+++ SGE/src/main/java/mocks/ReglaTemperaturaMayor20Grados.java
@@ -1,6 +1,6 @@
 package mocks;
 
-import regla.ReglaObservador;
+import reglas.ReglaObservador;
 
 public class ReglaTemperaturaMayor20Grados extends ReglaObservador {
 	
@@ -9,7 +9,7 @@ public class ReglaTemperaturaMayor20Grados extends ReglaObservador {
 	}
 
 	@Override
-	protected boolean seCumpleRegla(int temperaturaMedida) {
+	protected boolean seCumpleRegla(double temperaturaMedida) {
 		// Por ejemplo, definemos una regla para saber si la temperatura medida es mayor a 20 ºC
 		return temperaturaMedida > 20;
 	}
